+++ internal/models/user_test.go
@@ -168,13 +168,3 @@ func TestFromJSON(t *testing.T) {
 		t.Errorf("expected Version 2, got %d", user.Version)
 	}
 }
-
-// TestFromJSON_Invalid проверяет обработку невалидного JSON
-func TestFromJSON_Invalid(t *testing.T) {
-	invalidJSON := `{"id": "123", "login": ` // Оборванный JSON
-
-	_, err := models.FromJSON[models.User]([]byte(invalidJSON))
-	if err == nil {
-		t.Error("FromJSON() should return error for invalid JSON, but got nil")
-	}
-}
